BoxFolderInfoServiceのみDI化
