+++ src/main/java/api/box/service/impl/BoxApiConnectionService.java
@@ -1,4 +1,4 @@
-package api.box.service;
+package api.box.service.impl;
 
 import java.io.FileReader;
 import java.io.Reader;
