+++ src/main/java/api/box/service/impl/QuickStart.java
@@ -1,4 +1,4 @@
-package api.box.service;
+package api.box.service.impl;
 
 import java.io.BufferedReader;
 import java.io.FileReader;
